+++ src/boto3_assist/s3/__init__.py
@@ -1,0 +1,9 @@
+"""
+Geek Cafe, LLC
+Maintainers: Eric Wilson
+MIT License. See Project Root for the license information.
+"""
+
+from .s3 import S3
+
+__all__ = ["S3"]